requirements.txt
flask
celery
redis
yt-dlp
python-dotenv
werkzeug
Flask-Limiter
pydantic
cryptography
psycopg2-binary
sqlalchemy
alembic